refactor: Split MultiLayerNavPolicy into .cpp file
