+++ Core/include/Acts/Navigation/MultiLayerNavigationPolicy.hpp
@@ -48,24 +48,7 @@ class MultiLayerNavigationPolicy : public INavigationPolicy {
                                       const TrackingVolume& volume,
                                       const Logger& logger,
                                       const Config& config,
-                                      indexedUpdatorType grid)
-      : m_volume(volume), m_indexedGrid(std::move(grid)) {
-    ACTS_VERBOSE("Constructing MultiLayerNavigationPolicy for volume "
-                 << m_volume.volumeName());
-
-    // Fill the grid with surfaces
-    std::vector<std::shared_ptr<const Surface>> surfaces = {};
-    for (const auto& surface : m_volume.surfaces()) {
-      if (surface.associatedDetectorElement() == nullptr) {
-        continue;
-      }
-      surfaces.push_back(surface.getSharedPtr());
-    }
-
-    Experimental::detail::CenterReferenceGenerator rGenerator;
-    Experimental::detail::IndexedGridFiller filler{config.binExpansion};
-    filler.fill(gctx, m_indexedGrid, surfaces, rGenerator, {});
-  }
+                                      indexedUpdatorType grid);
 
   /// Update the navigation state from the surface array
   /// @param args The navigation arguments
@@ -73,35 +56,7 @@ class MultiLayerNavigationPolicy : public INavigationPolicy {
   /// @param logger The logger
   void initializeCandidates(const NavigationArguments& args,
                             AppendOnlyNavigationStream& stream,
-                            const Logger& logger) const {
-    ACTS_VERBOSE(
-        "MultiLayerNavigationPolicy Candidates initialization for volume"
-        << m_volume.volumeName());
-    const Transform3& itransform = m_volume.itransform();
-    const Vector3 locPosition = itransform * args.position;
-    const Vector3 locDirection = itransform * args.direction;
-
-    std::vector<Vector2> path = generatePath(locPosition, locDirection);
-
-    const auto& surfaces = m_volume.surfaces();
-    std::vector<const Surface*> surfCandidates = {};
-    surfCandidates.reserve(surfaces.size());
-
-    for (const auto& pos : path) {
-      std::vector<std::size_t> indices = m_indexedGrid.grid.atPosition(pos);
-
-      std::ranges::transform(indices, std::back_inserter(surfCandidates),
-                             [&](const auto& i) { return &surfaces[i]; });
-    }
-
-    ACTS_VERBOSE("MultiLayerNavigationPolicy Candidates reported"
-                 << surfCandidates.size() << " candidates");
-
-    // fill the navigation stream with the container
-    for (const auto* surf : surfCandidates) {
-      stream.addSurfaceCandidate(*surf, args.tolerance);
-    }
-  }
+                            const Logger& logger) const;
 
   /// Connect this policy with a navigation delegate
   /// @param delegate The navigation delegate to connect to
@@ -114,29 +69,7 @@ class MultiLayerNavigationPolicy : public INavigationPolicy {
   /// @param direction The direction of the path (in local frame)
   /// @return A vector of positions along the path
   std::vector<Vector2> generatePath(const Vector3& startPosition,
-                                    const Vector3& direction) const {
-    std::vector<Vector2> path;
-
-    auto maxXIndex = m_indexedGrid.grid.numLocalBins()[0];
-    auto maxYIndex = m_indexedGrid.grid.numLocalBins()[1];
-    Vector3 unitDir = direction.normalized();
-
-    for (std::size_t i = 0; i < maxYIndex; i++) {
-      auto v1 = m_indexedGrid.grid.lowerLeftBinEdge({1, i + 1});
-      auto v2 = m_indexedGrid.grid.upperRightBinEdge({maxXIndex, i + 1});
-
-      auto intersection = Acts::detail::IntersectionHelper2D::intersectSegment(
-          Vector2(v1[0], v1[1]), Vector2(v2[0], v2[1]),
-          startPosition.template block<2, 1>(0, 0),
-          unitDir.template block<2, 1>(0, 0));
-      if (!intersection.isValid()) {
-        continue;
-      }
-
-      path.push_back(intersection.position());
-    }
-    return path;
-  }
+                                    const Vector3& direction) const;
 
  private:
   // The tracking volume
