+++ Core/src/Navigation/CMakeLists.txt
@@ -5,4 +5,5 @@ target_sources(
         TryAllNavigationPolicy.cpp
         SurfaceArrayNavigationPolicy.cpp
         CylinderNavigationPolicy.cpp
+        MultiLayerNavigationPolicy.cpp
 )
